# docs/ai-tools.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ifc-edit-service",
    "version": "0.1.0"
  },
  "paths": {
    "/models/{id}/entities/{guid}": {
      "put": {
        "summary": "Put Entity",
        "description": "Apply edits to the in-memory model and record a pending change.",
        "operationId": "put_entity_models__id__entities__guid__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          },
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Put Entity Models  Id  Entities  Guid  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}/pending": {
      "get": {
        "summary": "Get Pending",
        "description": "List the current pending changes for a model.",
        "operationId": "get_pending_models__id__pending_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Pending Models  Id  Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Discard Pending",
        "description": "Discard pending changes: reload the in-memory model from disk.",
        "operationId": "discard_pending_models__id__pending_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Discard Pending Models  Id  Pending Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}/commit": {
      "post": {
        "summary": "Commit Pending",
        "description": "Atomically save all pending changes to disk and append them to history.\n\nThe first commit snapshots the original upload as ``v1`` before saving;\nevery commit snapshots the newly saved file as the next version. The\noptional body stamps ``operation`` onto the committed entries (default\n``update``; Go's override migration passes ``migrate``).",
        "operationId": "commit_pending_models__id__commit_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CommitBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Commit Pending Models  Id  Commit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}/history": {
      "get": {
        "summary": "Get History",
        "description": "List the persisted edit history for a model.",
        "operationId": "get_history_models__id__history_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get History Models  Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}/versions": {
      "get": {
        "summary": "Get Versions",
        "description": "List version snapshots for a model (empty + current=null before any commit).",
        "operationId": "get_versions_models__id__versions_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Versions Models  Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}/diff": {
      "post": {
        "summary": "Post Diff",
        "description": "Diff two model versions (or base version vs the current upload state).",
        "operationId": "post_diff_models__id__diff_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^m_[0-9a-f]{16}$",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post Diff Models  Id  Diff Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CommitBody": {
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "update",
              "migrate"
            ],
            "title": "Operation",
            "default": "update"
          }
        },
        "type": "object",
        "title": "CommitBody",
        "description": "Optional body of POST /models/{id}/commit."
      },
      "DiffBody": {
        "properties": {
          "base": {
            "type": "string",
            "title": "Base"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "base",
          "target"
        ],
        "title": "DiffBody",
        "description": "Body of POST /models/{id}/diff. target also accepts \"current\"."
      },
      "EditBody": {
        "properties": {
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "psets": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Psets"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": "local-user"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "type": "object",
        "title": "EditBody",
        "description": "Body of PUT /models/{id}/entities/{guid}."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Provenance": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "UI",
              "AI"
            ],
            "title": "Source",
            "default": "UI"
          }
        },
        "type": "object",
        "title": "Provenance",
        "description": "Who performed an edit: the web UI or an AI agent."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}